requirements.txt
fastapi
pydantic
uvicorn
motor
pymongo
python-jose
PyJWT
email_validator